Rename proposals_read_* files to proposals_*
This change makes the proposal-read module names consistent with the
circuit-read module names.
