+++ libsplinter/src/admin/rest_api/actix/mod.rs
@@ -17,8 +17,8 @@ pub(super) mod circuits;
 #[cfg(feature = "circuit-read")]
 pub(super) mod circuits_circuit_id;
 #[cfg(feature = "proposal-read")]
-pub(super) mod proposals_read;
+pub(super) mod proposals;
 #[cfg(feature = "proposal-read")]
-pub(super) mod proposals_read_circuit_id;
+pub(super) mod proposals_circuit_id;
 pub(super) mod submit;
 pub(super) mod ws_register_type;
